EndlessRuner - Test
I am taking a course about how to make an Endless Runer but I am too lazy to create a new Unity project and a new Github Repository.
The codes that I am going to learn can be used here so that's why I let them be on the same repository.